+++ src/koppeltaal/tests/test_definitions.py
@@ -41,7 +41,6 @@ def namespace():
 
 
 def test_coding(packer, namespace):
-
     vertebrates = koppeltaal.codes.Code(
         'Vertebrate',
         {'amphibians': 'Amphibians',