+++ src/koppeltaal/interfaces.py
@@ -109,10 +109,11 @@ class InvalidValue(KoppeltaalError):
         if self.field.interface is not None:
             python_name = self.field.interface.getName() + '.' + python_name
         if extended:
-            return "{}: invalid value for '{}' (FHIR name: '{}').".format(
-                self.__class__.__name__, python_name, self.field.name)
-        return "{}: invalid value for '{}'.".format(
-            self.__class__.__name__, python_name)
+            return "{}: invalid value '{}' for '{}' (FHIR name: '{}').".format(
+                self.__class__.__name__, self.value,
+                python_name, self.field.name)
+        return "{}: invalid value '{}' for '{}'.".format(
+            self.__class__.__name__, self.value, python_name)
 
 
 class InvalidReference(InvalidValue):