+++ CHANGES.md
@@ -6,6 +6,8 @@ Changes
 
 - Have the `InvalidValue` exception report a tiny bit more information.
 
+- Improvements for messages subcommand.
+
 1.3.5.3 (2018-11-16)
 --------------------